Add cli interface, moved to a separate module

+++ app/server.py
@@ -6,9 +6,8 @@ from random import randint
 from shutil import rmtree
 
 from flask import Flask, request, Response, abort, render_template
-from werkzeug.datastructures import FileStorage
-
 from pyvirtualdisplay import Display
+from werkzeug.datastructures import FileStorage
 
 Display().start()
 
@@ -20,12 +19,19 @@ app.debug = bool(os.environ.get("DEBUG"))
 logger = app.logger  # type: Logger
 
 temp_dir = Path(__file__).parent.parent.joinpath('var')
+
+try:
+    for _item in temp_dir.iterdir():
+        rmtree(str(_item))
+except FileNotFoundError:
+    pass
+
 temp_dir.mkdir(exist_ok=True)
 
 
-@app.route('/', methods=['GET'])
+@app.route('/', methods=['GET'], )
 def index_page():
-    return render_template('index.html')
+    return Response(render_template('index.html'), status=418)
 
 
 @app.route('/', methods=['POST'])
@@ -43,24 +49,25 @@ def generate():
 
     logger.info(f'Processing {rnd}')
 
-    temp_file = temp_dir.joinpath(f'archive_{rnd}.zip')
     temp_extracted = temp_dir.joinpath(rnd_dir)
     temp_extracted.mkdir()
-
-    temp_file.write_bytes(file.stream.read())
 
     try:
         templates = temp_extracted.absolute()
 
-        zipfile.ZipFile(temp_file).extractall(str(temp_extracted))
+        archive = zipfile.ZipFile(file)
+
+        if 'index.html' not in archive.namelist():
+            return abort(400, 'Index file not found')
+
+        archive.extractall(str(temp_extracted))
 
         data = process(f'file://{templates}/index.html')
 
         rmtree(str(temp_extracted))
-        temp_file.unlink()
 
         if data is None:
-            abort(500)
+            return abort(400, 'Failed to generate pdf')
 
         return Response(data, headers={
             'Content-Type': 'application/pdf',
@@ -69,10 +76,6 @@ def generate():
     except Exception as e:
         logger.error(e)
 
-        try:
-            temp_file.unlink()
-        except:
-            pass
         try:
             rmtree(str(temp_extracted))
         except:
